requirements.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
cryptography==36.0.1
h11==0.13.0
idna==3.3
joblib==1.1.0
lxml==4.8.0
numpy==1.22.3
outcome==1.1.0
pdfminer.six==20211012
pdfminer3k==1.3.4
pdfplumber==0.6.0
Pillow==9.0.1
ply==3.11
pycparser==2.21
pycryptodome==3.14.1
PyMySQL==1.0.2
pyOpenSSL==22.0.0
PyPDF2==1.26.0
PySocks==1.7.1
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
selenium==4.1.3
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
threadpoolctl==3.1.0
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.8
Wand==0.6.7
wsproto==1.1.0

pandas~=1.4.2
jieba~=0.42.1
matplotlib~=3.5.1
stylecloud~=0.5.2
bs4~=0.0.1
opencv-python~=4.7.0.72
wordcloud~=1.8.1
Appium-Python-Client~=2.11.1
SQLAlchemy~=1.4.36
redis~=4.3.4
Flask~=2.1.2